Fix the transfer buffer size
Previously the transfer buffer had an additional item
at the end to enforce the symmetry of the stencil. However,
having that item meant that in the situation of unmatched
grids the physical size of the interpolation region was
different in coarse and fine grids. This commit removes the
symmetry-enforcing item thus improving the performance of
transfers.